compartmentalize the sqlite firo cache code and add a background isolate worker to handle some processing as well as cache db writes
